+++ packages/scaffold-config/test/unit/supportFile.spec.ts
@@ -28,9 +28,6 @@ describe('supportFileComponent', () => {
           // Import commands.js using ES2015 syntax:
           import './commands'
 
-          // Alternatively you can use CommonJS syntax:
-          // require('./commands')
-
           import { mount } from '${mountModule}'
 
           Cypress.Commands.add('mount', mount)
@@ -61,9 +58,6 @@ describe('supportFileComponent', () => {
 
         // Import commands.js using ES2015 syntax:
         import './commands'
-
-        // Alternatively you can use CommonJS syntax:
-        // require('./commands')
 
         import { mount } from '${mountModule}'
 
@@ -112,9 +106,6 @@ describe('supportFileComponent', () => {
           // Import commands.js using ES2015 syntax:
           import './commands'
 
-          // Alternatively you can use CommonJS syntax:
-          // require('./commands')
-
           import { mount } from '${mountModule}'
 
           Cypress.Commands.add('mount', mount)
@@ -145,9 +136,6 @@ describe('supportFileComponent', () => {
 
         // Import commands.js using ES2015 syntax:
         import './commands'
-
-        // Alternatively you can use CommonJS syntax:
-        // require('./commands')
 
         import { mount } from '${mountModule}'
 
@@ -195,9 +183,6 @@ describe('supportFileComponent', () => {
         // Import commands.js using ES2015 syntax:
         import './commands'
 
-        // Alternatively you can use CommonJS syntax:
-        // require('./commands')
-
         import { mount } from '${mountModule}'
 
         // Augment the Cypress namespace to include type definitions for
@@ -244,9 +229,6 @@ describe('supportFileComponent', () => {
           // Import commands.js using ES2015 syntax:
           import './commands'
 
-          // Alternatively you can use CommonJS syntax:
-          // require('./commands')
-
           import { mount } from 'cypress/svelte'
 
           Cypress.Commands.add('mount', mount)
@@ -277,9 +259,6 @@ describe('supportFileComponent', () => {
 
         // Import commands.js using ES2015 syntax:
         import './commands'
-
-        // Alternatively you can use CommonJS syntax:
-        // require('./commands')
 
         import { mount } from 'cypress/svelte'