+++ system-tests/project-fixtures/svelte/cypress/support/component.js
@@ -1,4 +1,0 @@
-import { mount } from 'cypress/svelte'
-import '../../src/styles.css'
-
-Cypress.Commands.add('mount', mount)