+++ system-tests/project-fixtures/svelte/src/App.cy.ts
@@ -1,7 +1,11 @@
 import App from './App.svelte'
 
 it('should render with style', () => {
-  cy.mount(App)
+  cy.mount(App, {
+    // logging due to @packages/app/reporter-ct-mount-hover.cy.ts tests to test the mount log
+    log: true,
+  })
+
   cy.contains('Hello World!')
   // Verify global styles
   cy.get('.very-red').should('have.css', 'color', 'rgb(255, 0, 0)')