+++ system-tests/projects/angular-cli-configured/cypress/support/component.ts
@@ -13,8 +13,8 @@
 // https://on.cypress.io/configuration
 // ***********************************************************
 
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
+// Import commands.js using ES2015 syntax:
+import './commands'
 
 import { mount } from 'cypress/angular'