+++ system-tests/projects/no-specs-vue/cypress/support/component.ts
@@ -16,9 +16,6 @@
 // Import commands.js using ES2015 syntax:
 import './commands'
 
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
-
 // Import global styles
 import '@/assets/main.css'