+++ system-tests/test-binary/node_versions_spec.ts
@@ -31,6 +31,7 @@ describe('binary node versions', () => {
     'cypress/base:20.18.0',
     'cypress/base:22.0.0',
     'cypress/base:22.7.0',
+    'cypress/base:22.12.0',
   ].forEach(smokeTestDockerImage)
 })
 
@@ -41,6 +42,7 @@ describe('type: module', () => {
     'cypress/base:20.18.0',
     'cypress/base:22.0.0',
     'cypress/base:22.7.0',
+    'cypress/base:22.12.0',
   ].forEach((dockerImage) => {
     systemTests.it(`can run in ${dockerImage}`, {
       withBinary: true,